Consolidate Logics for GPU Detection
This PR unifies automatic device detection logic by using
`mlc_chat.support.auto_device`, which comes with detailed logging and
fallback mechanisms.
